removed comments and printouts

+++ ProjectDico/MNCalendarViewDayCell.m
@@ -2,7 +2,7 @@
 //  MNCalendarViewDayCell.m
 //  ProjectDico
 //
-//  Created by Nicole Zhu on 5/25/14.
+//  Created by Nicole Zhu on 6/02/14.
 //  Copyright (c) 2014 ProjectDico. All rights reserved.
 //
 
@@ -37,7 +37,7 @@ NSString *const MNCalendarViewDayCellIdentifier = @"MNCalendarViewDayCellIdentif
                      fromDate:self.month];
     
     self.weekday = components.weekday;
-    self.titleLabel.text = [NSString stringWithFormat:@"%d", components.day];
+    self.titleLabel.text = [NSString stringWithFormat:@"%ld",(long)components.day];
     self.enabled = monthComponents.month == components.month;
     self.currentWeek = NO;
